refactor: eliminate redundant business address fields
The Business Address fields we have in SST are redundant: Users are
forced to enter the same exact addresses in their TaxCloud account
and on the SST settings page. This makes for a poor UX. On top of
that, there are some serious issues with the way we currently save
origin addresses for products. We save the _index_ of the selected
origin addresses instead of a real unique identifier, which means
if the order of the addresses entered on the settings page changes
then the origin address for one or more products could also change!

This commit includes changes to address both of these flaws in the
current design. First, it eliminates the redundant Business Address
settings in favor of pulling the locations from the merchant's
TaxCloud account with the GetLocations API. Second, it refactors the
code to use TaxCloud's LocationID as the key for each origin address
instead of the address's index.

An update routine is included to handle importing the merchant's
addresses from TaxCloud and updating from the old schema to the new
schema for saving product origin addresses. If any problems are
detected during the update a prominent notice will be displayed to
warn the admin. The probability of this happening should be very
small.

closes #55