refactor: use 'key' property to identify cart items
We assume that the cart item's unique identifier is always used as the key in
the `contents` array, but some plugins like WCFM transform the WC cart packages
in a way that violates this assumption. By ensuring that a 'key' property with
the unique ID of the item always exists and using it instead of the array key
to identify items, we can maximize compatability with third parties like WCFM
that might transform the `contents` array in unexpected ways.