fix: refresh settings values before setting an option
Currently, we do not always read the latest settings values from the
options table before setting a new option value in SST_Settings::set().
This means that calls to SST_Settings::set() can wind up overwriting
settings values set earlier in the request, which was the root cause
of #67.

fixes #67