fix: prevent customers outside of US from adding exemption certificates
Show a message "Billing address must be in the United States to add an
exemption certificate" if a customer outside of the US clicks the
"Add Certificate" button.

The TaxCloud API can only accept certificates from customers with an
address in the US, so it doesn't make sense to allow customers located
elsewhere to open the certificate form.

fixes #116